requirements.txt
annotated-types==0.7.0
certifi==2024.7.4
influxdb-client==1.45.0
paho-mqtt==2.1.0
pydantic==2.8.2
pydantic_core==2.20.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
reactivex==4.0.4
six==1.16.0
typing_extensions==4.12.2
urllib3==2.2.2
